add back v1 syntax export for symbols
The introduction of the _NiceJsonContext helper class is in response to
tweaks to bring back the trailing newline that I had apparently lost
in the update to v2 files.
The indent_step arg seemed really annoying to pass around, and I
wanted different argument semantics internally vs externally.
(final_newline makes no sense for recursive calls yet should be enabled
by default for external calls. 'indent' meaning 'indentation step' is
best for the external interface to be consistent with the json module,
but in the implementation it makes more sense for 'indent' to mean
the current indent)